# perf: Use custom allocator