@charliermarsh 2023-02-11
Nice! Are there any costs or risks associated with this apart from the increased binary size? Have you used custom allocators in previous projects? I've seen them mentioned in [Nethercote's book](https://nnethercote.github.io/perf-book/heap-allocations.html#using-an-alternative-allocator) but never tried them :)