@MichaReiser 2023-02-11
The book you mention summarises it well

> The exact effect will depend on the individual program and the alternative allocator chosen, but large improvements in speed and very large reductions in memory usage have been seen in practice. The effect will also vary across platforms, because each platform’s system allocator has its own strengths and weaknesses. The use of an alternative allocator can also affect binary size.

We could try to measure peak memory usage and compare it between allocators. 

The main risk I see is that it introduces an additional surface that needs maintaining. We need to update the allocator when it has a known security issue and an it may be more likely that a custom allocator has bugs which I would expect less from the system allocators.