@MichaReiser 2023-02-11
I'm undecided whether the performance improvement on Linux warrants such a significant binary size increase.

But, using jemalloc on all unix platforms has a harmonizing effect, reducing the differences between the two platforms.