@andersk 2023-02-11
Much of the size difference on Linux is due to debug symbols. (NixOS 23.05 on Ryzen 7 1800X, averages of 100 runs.)

allocator | size | stripped | cpython time | cpython RSS memory
--- | --- | --- | --- | ---
system | 13.01 MiB | 9.32 MiB | 407 ms | 131 MiB
jemalloc | 18.25 MiB | 9.92 MiB | 378 ms | 175 MiB
mimalloc | 13.04 MiB | 9.37 MiB | 454 ms | 202 MiB
mimalloc (no `secure`) | 13.03 MiB | 9.36 MiB | 401 ms | 197 MiB