@andersk 2023-02-11
The memory use change is a bit concerning though—with Ruff already being as blindingly fast as it is, is a 7% speed increase really worth a 33% memory use increase?