@charliermarsh 2023-02-11
Ahhh that's really interesting, thank you for running + tabulating. And indeed that's a bit concerning. I was going to say that perhaps it makes sense to ship on Windows but not elsewhere given that the speed increase in the PR summary is much more significant, but presumedly (?) that would result in a similar memory use increase as the `mimalloc` benchmarks above.