@MichaReiser 2023-02-12
@andersk, what command did you use to get the peak memory usage? 

I used mac's instruments to get some numbers before and after.

<table>
<tr>
	<td>system allocator
	<td>jemalloc
<tr>
	<td><img width="1792" alt="Memory usage with system allocator" src="https://user-images.githubusercontent.com/1203881/218298010-0ed86279-b2d4-42a3-9876-2d8bb36a84b8.png"> 
	<td><img width="1792" alt="Memory usage with jemalloc allocator" src="https://user-images.githubusercontent.com/1203881/218298026-244b4f99-ae8f-4107-97d3-32f804bb66db.png">
<tr>
	<td>Peak: 42MB
	<td>Peak: 145MB
</table>

This looks very concerning but these measurements aren't necessarily fair. My understanding of the numbers shown by Instruments is that it shows `total(malloc) - total(free)` over time. This is useful for understanding the allocations in your program but only has limited meaning in assessing how much physical memory your application uses because it doesn't account for heap fragmentation and the fact that the OS allocates the memory in pages (~4Kib). Meaning, you can have a situation where the OS isn't able to free a single page because your application has freed all allocations except one on every page. 

Using `jemalloc` gives you an entirely different graph because `jemalloc` handles the memory pages manually. It is an abstraction on top of `malloc` and `free`. That's why the graph only shows a few but larger allocations. Now, why does `jemalloc` not return memory? One possibility is that the heap is too fragmented to release a single page or `ruff` terminates before `jemalloc` runs its memory release routine. 

> This was a new notion to me. Surely a good piece of code, which behaves itself and puts away its toys when it's done, will keep its resource usage to a minimum? As it turns out, that isn't exactly the case. It really comes down to the behavior of the allocator, and most allocators won't necessarily return memory right away to the OS. Jemalloc has a few ways this can be tuned, but by default it uses a time-based strategy for returning memory back to the OS. Memory allocations are serviced first by memory already held by the process that is free for reuse, and then failing that, new memory is acquired from the OS. Given enough time, should a chunk of memory not have anything assigned to it, it will be returned (and if returning resources quickly is important to you, you can configure jemalloc to return the memory immediately).
[source](https://engineering.linkedin.com/blog/2021/taming-memory-fragmentation-in-venice-with-jemalloc)

I recommend investing time to find a long-running workload and tooling that measures the reserved OS memory if we're concerned about peak memory consumption. I myself aren't too concerned, considering that 150MB for lining the whole cpython code seems reasonable. 

> @MichaReiser - Should we consider putting this behind a flag while we continue to discuss, based on the above?

I personally would probably just revert the PR and reopen it. It's not a lot of code. 

For JS tooling it is very common to use custom allocators:
* [SWC](https://github.com/swc-project/swc/pull/4791)
* [lightning_css](https://github.com/parcel-bundler/lightningcss/commit/18adb3be5215c74b5dfa9a495c0a11e959e0d361)
* [Rome](https://github.com/rome/tools/pull/2900)
* [Oxc](https://github.com/Boshen/oxc/pull/12)