@andersk 2023-02-12
The `time(1)` command on Linux gives you memory information, but you have to make sure to ask for the command and not the shell builtin:

```console
$ time ruff -n ../cpython
…
real    0m0.710s
user    0m5.472s
sys     0m0.146s

$ command time ruff -n ../cpython
…
5.39user 0.15system 0:00.79elapsed 695%CPU (0avgtext+0avgdata 112264maxresident)k
0inputs+0outputs (0major+28331minor)pagefaults 0swaps

$ \time ruff -n ../cpython
…
5.47user 0.15system 0:00.71elapsed 786%CPU (0avgtext+0avgdata 111580maxresident)k
0inputs+0outputs (0major+28573minor)pagefaults 0swaps
```